+++ DuetRRFPlugin.py
@@ -1,148 +1,119 @@
-import re
-import os.path
+import os
 import json
-
-from PyQt5.QtCore import QObject, QUrl, pyqtProperty, pyqtSignal, pyqtSlot
-from PyQt5.QtQml import QQmlComponent, QQmlContext
-
-from UM.Message import Message
-from UM.Logger import Logger
-
-from UM.Extension import Extension
-from UM.PluginRegistry import PluginRegistry
-from UM.OutputDevice.OutputDevicePlugin import OutputDevicePlugin
-
-from . import DuetRRFOutputDevice
-from UM.i18n import i18nCatalog
-catalog = i18nCatalog("cura")
+import re
 
 from cura.CuraApplication import CuraApplication
 
+from UM.Message import Message
+from UM.Logger import Logger
+from UM.Extension import Extension
+from UM.PluginRegistry import PluginRegistry
+from UM.OutputDevice.OutputDevicePlugin import OutputDevicePlugin
+from UM.i18n import i18nCatalog
+catalog = i18nCatalog("cura")
 
-class DuetRRFPlugin(QObject, Extension, OutputDevicePlugin):
-    def __init__(self, parent=None):
-        QObject.__init__(self, parent)
-        Extension.__init__(self)
-        OutputDevicePlugin.__init__(self)
-        self.addMenuItem(catalog.i18n("DuetRRF Connections"), self.showSettingsDialog)
-        self._dialogs = {}
-        self._dialogView = None
+from .DuetRRFOutputDevice import DuetRRFOutputDevice, DuetRRFDeviceType
 
-        CuraApplication.getInstance().getPreferences().addPreference("duetrrf/instances", json.dumps({}))
-        self._instances = json.loads(CuraApplication.getInstance().getPreferences().getValue("duetrrf/instances"))
 
-    def start(self):
-        manager = self.getOutputDeviceManager()
-        for name, instance in self._instances.items():
-            manager.addOutputDevice(DuetRRFOutputDevice.DuetRRFOutputDevice(name, instance["url"], instance["duet_password"], instance["http_user"], instance["http_password"], device_type=DuetRRFOutputDevice.DeviceType.print))
-            manager.addOutputDevice(DuetRRFOutputDevice.DuetRRFOutputDevice(name, instance["url"], instance["duet_password"], instance["http_user"], instance["http_password"], device_type=DuetRRFOutputDevice.DeviceType.simulate))
-            manager.addOutputDevice(DuetRRFOutputDevice.DuetRRFOutputDevice(name, instance["url"], instance["duet_password"], instance["http_user"], instance["http_password"], device_type=DuetRRFOutputDevice.DeviceType.upload))
+class DuetRRFPlugin(Extension, OutputDevicePlugin):
+    def __init__(self):
+        super().__init__()
+        self._application = CuraApplication.getInstance()
+        self._application.globalContainerStackChanged.connect(self.checkDuetRRFOutputDevices)
 
-    def stop(self):
-        manager = self.getOutputDeviceManager()
-        for name in self._instances.keys():
-            manager.removeOutputDevice(name + "-print")
-            manager.removeOutputDevice(name + "-simulate")
-            manager.removeOutputDevice(name + "-upload")
+        # LEGACY code - remove me!
+        # only add legacy menu item if legacy settings exist
+        try:
+            self._instances = json.loads(CuraApplication.getInstance().getPreferences().getValue("duetrrf/instances"))
+            Logger.log("d", "found legacy settings for {}".format(", ".join(self._instances.keys())))
+            self.addMenuItem(catalog.i18n("(moved to Preferences→Printer)"), self.showSettingsDialog)
+            self.addMenuItem(catalog.i18n("(delete all old legacy settings)"), self.deleteAllOldLegacySettings)
+        except:
+            pass
 
-    def _createDialog(self, qml):
-        path = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'resources', 'qml', qml)
-        dialog = CuraApplication.getInstance().createQmlComponent(path, {"manager": self})
-        return dialog
+    def migrateOldSettings(self):
+        """
+        This is only to provide legacy compatibility and will be removed in a future release.
+        """
+        try:
+            self._instances = json.loads(CuraApplication.getInstance().getPreferences().getValue("duetrrf/instances"))
+        except:
+            return
 
-    def _showDialog(self, qml):
-        if not qml in self._dialogs:
-            self._dialogs[qml] = self._createDialog(qml)
-        self._dialogs[qml].show()
+        global_container_stack = self._application.getGlobalContainerStack()
+        name = global_container_stack.getName()
+
+        if name in self._instances.keys():
+            Logger.log("d", "migrating legacy settings for {}".format(name))
+            global_container_stack.setMetaDataEntry("duetrrf", True)
+            global_container_stack.setMetaDataEntry("duetrrf_url", self._instances[name]["url"])
+            global_container_stack.setMetaDataEntry("duetrrf_duet_password", self._instances[name]["duet_password"])
+            global_container_stack.setMetaDataEntry("duetrrf_http_user", self._instances[name]["http_user"])
+            global_container_stack.setMetaDataEntry("duetrrf_http_password", self._instances[name]["http_password"])
+
+            # delete this printer from the old settings - the new metadata is now the only source of truth.
+            del self._instances[name]
+            if len(self._instances) == 0:
+                CuraApplication.getInstance().getPreferences().removePreference("duetrrf/instances")
+            else:
+                CuraApplication.getInstance().getPreferences().setValue("duetrrf/instances", json.dumps(self._instances))
 
     def showSettingsDialog(self):
-        self._showDialog("DuetRRFPlugin.qml")
+        """
+        This is only to provide legacy compatibility and will be removed in a future release.
+        """
+        Logger.log("d", "DuetRRF showSettingsDialog called.")
+        message = Message(
+            "Settings for the DuetRRF plugin moved to the Printer preferences.\n\n"
+            "Please go to:\n"
+            "\t → Cura Preferences\n"
+            "\t → Printers\n"
+            "\t → activate and select your printer\n"
+            "\t → click on 'Connect Duet RepRapFirmware'\n\n"
+            "You still have legacy settings for other printers:\n{}\n\n"
+            "Please check all your printer settings!".format(
+                ",\n".join(self._instances.keys())
+            ),
+            lifetime=0,
+            title="Deprecation: DuetRRF settings have moved to Cura Preferences!",
+        )
+        message.show()
 
-    serverListChanged = pyqtSignal()
-    @pyqtProperty("QVariantList", notify=serverListChanged)
-    def serverList(self):
-        return list(self._instances.keys())
+    def deleteAllOldLegacySettings(self):
+        """
+        This is only to provide legacy compatibility and will be removed in a future release.
+        """
+        Logger.log("d", "DuetRRF deleteAllOldLegacySettings called.")
+        CuraApplication.getInstance().getPreferences().removePreference("duetrrf/instances")
+        message = Message(
+            "Legacy settings have been deleted for the following printers:\n{}\n\n"
+            "Please restart Cura.".format(
+                ",\n".join(self._instances.keys())
+            ),
+            lifetime=0,
+            title="Legacy DuetRRF settings successfully deleted!",
+        )
+        message.show()
 
-    @pyqtSlot(str, result=str)
-    def instanceUrl(self, name):
-        if name in self._instances.keys():
-            return self._instances[name]["url"]
-        return None
+    def checkDuetRRFOutputDevices(self):
+        global_container_stack = self._application.getGlobalContainerStack()
+        if not global_container_stack:
+            return
 
-    @pyqtSlot(str, result=str)
-    def instanceDuetPassword(self, name):
-        if name in self._instances.keys():
-            return self._instances[name]["duet_password"]
-        return None
+        self.migrateOldSettings() # LEGACY code - remove me!
 
-    @pyqtSlot(str, result=str)
-    def instanceHTTPUser(self, name):
-        if name in self._instances.keys():
-            return self._instances[name]["http_user"]
-        return None
-
-    @pyqtSlot(str, result=str)
-    def instanceHTTPPassword(self, name):
-        if name in self._instances.keys():
-            return self._instances[name]["http_password"]
-        return None
-
-    @pyqtSlot(str, str, str, str, str, str)
-    def saveInstance(self, oldName, name, url, duet_password, http_user, http_password):
-        if oldName:
-            # this is a edit operation, delete the old instance before saving the new one
-            self.removeInstance(oldName)
-
-        if not url.endswith('/'):
-            url += '/'
-
-        self._instances[name] = {
-            "url": url,
-            "duet_password": duet_password,
-            "http_user": http_user,
-            "http_password": http_password
-        }
         manager = self.getOutputDeviceManager()
-        manager.addOutputDevice(DuetRRFOutputDevice.DuetRRFOutputDevice(name, url, duet_password, http_user, http_password, device_type=DuetRRFOutputDevice.DeviceType.print))
-        manager.addOutputDevice(DuetRRFOutputDevice.DuetRRFOutputDevice(name, url, duet_password, http_user, http_password, device_type=DuetRRFOutputDevice.DeviceType.simulate))
-        manager.addOutputDevice(DuetRRFOutputDevice.DuetRRFOutputDevice(name, url, duet_password, http_user, http_password, device_type=DuetRRFOutputDevice.DeviceType.upload))
-        CuraApplication.getInstance().getPreferences().setValue("duetrrf/instances", json.dumps(self._instances))
-        self.serverListChanged.emit()
-        Logger.log("d", "Instance saved: " + name)
 
-    @pyqtSlot(str)
-    def removeInstance(self, name):
-        manager = self.getOutputDeviceManager()
-        manager.removeOutputDevice(name + "-print")
-        manager.removeOutputDevice(name + "-simulate")
-        manager.removeOutputDevice(name + "-upload")
-        del self._instances[name]
-        CuraApplication.getInstance().getPreferences().setValue("duetrrf/instances", json.dumps(self._instances))
-        self.serverListChanged.emit()
-        Logger.log("d", "Instance removed: " + name)
+        # remove all DuetRRF output devices - the new stack might not need them
+        manager.removeOutputDevice("duetrrf-print")
+        manager.removeOutputDevice("duetrrf-simulate")
+        manager.removeOutputDevice("duetrrf-upload")
 
-    @pyqtSlot(str, str, result = bool)
-    def validName(self, oldName, newName):
-        if not newName:
-            # empty string isn't allowed
-            return False
-        if oldName == newName:
-            # if name hasn't changed, it is not a duplicate
-            return True
-
-        # duplicates not allowed
-        return (not newName in self._instances.keys())
-
-    @pyqtSlot(str, str, result = bool)
-    def validUrl(self, oldName, newUrl):
-        if newUrl.startswith('\\\\'):
-            # no UNC paths
-            return False
-        if not re.match('^https?://.', newUrl):
-            # missing https?://
-            return False
-        if '@' in newUrl:
-            # @ is probably HTTP basic auth, which is a separate setting
-            return False
-
-        return True
+        # check and load new output devices
+        if global_container_stack.getMetaDataEntry("duetrrf"):
+            Logger.log("d", "DuetRRF is active for printer: " + global_container_stack.getName())
+            manager.addOutputDevice(DuetRRFOutputDevice("duetrrf-print", DuetRRFDeviceType.print))
+            manager.addOutputDevice(DuetRRFOutputDevice("duetrrf-simulate", DuetRRFDeviceType.simulate))
+            manager.addOutputDevice(DuetRRFOutputDevice("duetrrf-upload", DuetRRFDeviceType.upload))
+        else:
+            Logger.log("d", "DuetRRF is not active for printer: " + global_container_stack.getName())