+++ __init__.py
@@ -1,7 +1,9 @@
-from . import DuetRRFPlugin
+import os
+import json
 
-from UM.i18n import i18nCatalog
-catalog = i18nCatalog("cura")
+from UM.Logger import Logger
+
+from . import DuetRRFPlugin, DuetRRFAction
 
 
 def getMetaData():
@@ -9,8 +11,17 @@ def getMetaData():
 
 
 def register(app):
+    plugin_file_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "plugin.json")
+    try:
+        with open(plugin_file_path) as plugin_file:
+            plugin_info = json.load(plugin_file)
+            Logger.log("d", "DuetRRFPlugin version: {}".format(plugin_info["version"]))
+    except:
+        Logger.log("w", "DuetRRFPlugin failed to get version information!")
+
     plugin = DuetRRFPlugin.DuetRRFPlugin()
     return {
         "extension": plugin,
-        "output_device": plugin
+        "output_device": plugin,
+        "machine_action": DuetRRFAction.DuetRRFAction()
     }